+++ modules/admin/src/models/StorageFilter.php
@@ -113,7 +113,7 @@ class StorageFilter extends \admin\ngrest\base\Model
 
     // ngrest
 
-    public function ngRestApiEndpoint()
+    public static function ngRestApiEndpoint()
     {
         return 'api-admin-filter';
     }