+++ modules/admin/src/ngrest/interfaces/ConfigBuilder.php
@@ -5,6 +5,7 @@ namespace admin\ngrest\interfaces;
 /**
  * Ngrest Config Builder Interface
  * 
+ * @todo rename to ConfigBuilderInterface
  * @author Basil Suter <basil@nadar.io>
  */
 interface ConfigBuilder