+++ modules/crawleradmin/models/Builderindex.php
@@ -94,7 +94,7 @@ class Builderindex extends \admin\ngrest\base\Model
         return ['url', 'content', 'title', 'language_info'];
     }
 
-    public function ngRestApiEndpoint()
+    public static function ngRestApiEndpoint()
     {
         return 'api-crawler-builderindex';
     }