+++ modules/crawleradmin/models/Searchdata.php
@@ -66,7 +66,7 @@ class Searchdata extends \admin\ngrest\base\Model
     /**
      * @return string Defines the api endpoint for the angular calls
      */
-    public function ngRestApiEndpoint()
+    public static function ngRestApiEndpoint()
     {
         return 'api-crawler-searchdata';
     }