+++ modules/newsadmin/models/Tag.php
@@ -40,7 +40,7 @@ class Tag extends \admin\ngrest\base\Model
 
     public $i18n = ['title'];
 
-    public function ngRestApiEndpoint()
+    public static function ngRestApiEndpoint()
     {
         return 'api-news-tag';
     }